fix: 게시글 수정 버그 수정 -db상에서 삭제안되고 aws에서만 삭제되던 부분 수정 - 새로운 url만 따로 첨부파일 또는 이미지 파일 db에 추가 - 없어진 파일은 제거 - 원래 있던파일에 대해선 아무것도 안함 #42

+++ src/main/java/github/com/jbabe/service/gallery/GalleryService.java
@@ -45,7 +45,11 @@ public class GalleryService {
         if(pageable.getPageNumber()+1>galleryPages.getTotalPages()) throw new NotFoundException("Page Not Found", pageable.getPageNumber());
 
         for(Gallery gallery: galleryPages){
-            GalleryImg firstImg = gallery.getGalleryImgs().get(0);
+
+            List<GalleryImg> galleryImgs = gallery.getGalleryImgs();
+            GalleryImg firstImg = galleryImgs.isEmpty()?
+                    GalleryImg.builder().fileName("갤러리 없는 갤러리 게시물").fileUrl("https://www.irisoele.com/img/noimage.png").build():
+                    gallery.getGalleryImgs().get(0);
 
             GalleryListDto galleryListDto = GalleryMapper.INSTANCE
                     .GalleryToGalleryListDto(gallery, firstImg.getFileName(), firstImg.getFileUrl());
@@ -102,13 +106,16 @@ public class GalleryService {
     }
 
     @Transactional
-    public void modifyGalleryPost(int galleryId, GalleryDetailsDto requestModify, boolean isOfficial){
+    public void modifyGalleryPost(int galleryId, GalleryDetailsDto requestModify, Boolean isOfficial){
         Gallery orginalGallery = galleryJpa.findById(galleryId).orElseThrow(
                 ()-> new NotFoundException("Not Found Gallery", galleryId));
         List<GalleryImg> originalImgs = orginalGallery.getGalleryImgs();
 
         List<GalleryImg> imagesToBeErased = updateAndRemoveNonMatching(originalImgs, requestModify.getFiles());
-        galleryImgJpa.deleteAll(imagesToBeErased);
+        if (!imagesToBeErased.isEmpty()) {
+            orginalGallery.getGalleryImgs().removeAll(imagesToBeErased);
+            galleryImgJpa.deleteAll(imagesToBeErased);
+        }
 
         List<GalleryImg> imageToBeAdded = forMissingFilesClickAdd(requestModify.getFiles(), orginalGallery);
         orginalGallery.notifyAndEditSubjectLineContent(requestModify, isOfficial, imageToBeAdded);
